Fixed two bugs in SimpleWeb and upgraded client timeout pinging
The first bug is inside SimpleWeb.jslib where a hanging }); prevented
unity from building.

The second bug is inside Connection.cs where if the client had
X-Forwarded-For in their headers the server attempted to parse the port
but failed.

The SimpleWeb.jslib file was updated to use a Web Worker instead of
setInterval for a more robust background-thread safe way of pinging the
server and keeping the connection alive.